Change int on size_t in insert_spaced function

+++ src/appleseed.studio/mainwindow/pythoninput.h
@@ -55,7 +55,7 @@ class PythonInput
   private:
     void indent();
     void indent_like_previous(const std::string &previous);
-    void insert_spaces(int count);
+    void insert_spaces(size_t count);
 };
 
 }       // namespace studio
